feat: 상품 조회 기능 구현

+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductData.java
@@ -5,7 +5,7 @@ import com.loopers.support.error.CoreException;
 import com.loopers.support.error.ErrorType;
 
 
-public record ProductData(String productName, int amount, int quantity, String brandName, int likeCount) {
+public record ProductData(String productName, int amount, int quantity, String brandName, long likeCount) {
     public static ProductData from(Product product, Brand brand, int likeCount) {
         if (product == null || brand == null) {
             throw new CoreException(ErrorType.NOT_FOUND, "상품 또는 브랜드 정보가 누락되었습니다.");
